SocialConnect/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.4
django-crispy-forms==2.0
mysqlclient==2.2.0
sqlparse==0.4.4
typing-extensions==4.7.1
